brainbox/release/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kaia-brainbox"
description = "Web-server and API to access the curated set of AI-deciders, packaged as containers"
version = "0.0.6"
readme = 'README.md'
requires-python = ">=3.10, <4.0"

license = { text = "GPLv3" }

authors = [
    {name = "Yuri Okulovsky", email = "yuri.okulovsky@gmail.com"}
]

dependencies = [
    "yo_fluq_ds",
    "flask",
    "requests",
    "sqlalchemy"
]

classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Information Technology",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Development Status :: 4 - Beta"
]

[project.optional-dependencies]
dev = [
    "tox",
    "tox-conda",
    "twine"
]

[project.urls]
repository = "https://github.com/okulovsky/kaia/tree/main/brainbox"

[tool.setuptools.packages.find]
where = ["."]
